genesis_bots/genesis_sample_golden/original_teams_justin/requirements.txt
flask==3.0.0
botbuilder-core==4.14.0
botbuilder-schema==4.14.0
botframework-connector==4.14.0
aiohttp==3.7.4
cryptography>=3.0
requests>=2.0.0
typing-extensions ; python_version < "3.5.2"